e2e/requirements.txt
pytest>=8.0,<9
pytest-playwright>=0.5,<1
pytest-base-url>=2.1,<3
